+++ component/async/kafka/option.go
@@ -95,6 +95,7 @@ func DecoderJSON() OptionFunc {
 // It requires a function indicating how to extract the time from a Kafka message.
 func WithDurationOffset(since time.Duration, timeExtractor TimeExtractor) OptionFunc {
 	return func(c *ConsumerConfig) error {
+		c.DurationBasedConsumer = true
 		c.DurationOffset = since
 		c.TimeExtractor = timeExtractor
 		return nil
